[SPARK-46504][PS][TESTS][FOLLOWUPS] Moving slow tests out of `IndexesTests`: conversion and drop methods
### What changes were proposed in this pull request?
Moving slow tests out of `IndexesTests`: conversion and drop methods

### Why are the changes needed?
for testing parallelism

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
updated ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44536 from zhengruifeng/ps_test_idx_base_conversion_drop.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
